CHORE(test): remove duplicate/skipped tests
This commit collapses the duplication in account testing introduced in
 #30.
